restructure favorites page
