+++ pages/artworks-collection/index.js
@@ -1,7 +1,6 @@
 import ArtworksList from "@/components/ArtworksList/index.js";
 import Navigation from "@/components/Navigation";
 import SearchBar from "@/components/SearchBar";
-import { styled } from "styled-components";
 
 export default function ArtworkCollection({
   artworks,
@@ -23,7 +22,3 @@ export default function ArtworkCollection({
     </>
   );
 }
-
-const StyledButton = styled.button`
-  margin-bottom: 80px;
-`;
